Fix DataConverter inconsistencies
